state held in app.js
