+++ frontend/src/assets/css/01-base/_fonts.sass
@@ -1,1 +1,1 @@
-@import url('https://fonts.googleapis.com/css?family=Fira+Sans:300,400,400i')
+@import url('https://fonts.googleapis.com/css?family=Fira+Sans:300,400');
